Rename from "loader" to "launcher"
