+++ lib/windows/app/actions/mainMenuActions.js
@@ -36,17 +36,17 @@
 
 import { ipcRenderer } from 'electron';
 
-export const OPEN_APP_LOADER = 'MAIN_MENU_OPEN_APP_LOADER';
+export const OPEN_APP_LAUNCHER = 'MAIN_MENU_OPEN_APP_LAUNCHER';
 
-function openAppLoaderAction() {
+function openAppLauncherAction() {
     return {
-        type: OPEN_APP_LOADER,
+        type: OPEN_APP_LAUNCHER,
     };
 }
 
-export function openAppLoader() {
+export function openAppLauncher() {
     return dispatch => {
-        dispatch(openAppLoaderAction());
-        ipcRenderer.send('open-app-loader');
+        dispatch(openAppLauncherAction());
+        ipcRenderer.send('open-app-launcher');
     };
 }
