+++ lib/windows/app/containers/MainMenuContainer.js
@@ -47,9 +47,9 @@ function mapDispatchToProps(dispatch) {
     return {
         menuItems: [{
             id: 1,
-            text: 'Load other app...',
+            text: 'Launch other app...',
             hotkey: 'Alt+L',
-            onClick: () => dispatch(MainMenuActions.openAppLoader()),
+            onClick: () => dispatch(MainMenuActions.openAppLauncher()),
         }],
     };
 }
