+++ lib/windows/launcher/actions/appsActions.js
@@ -34,27 +34,61 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { Record } from 'immutable';
-import * as NavMenuActions from '../actions/navMenuActions';
+import { ipcRenderer } from 'electron';
+import { getInstalledApps } from '../../../util/apps';
+import * as ErrorDialogActions from '../../../actions/errorDialogActions';
 
-const InitialState = Record({
-    menuItems: [
-        { id: 1, text: 'Load app', iconClass: 'icon-ellipsis-vert' },
-        { id: 2, text: 'Add/remove apps', iconClass: 'icon-list-add' },
-        { id: 3, text: 'Settings', iconClass: 'icon-wrench' },
-    ],
-    selectedItemId: 1,
-});
+export const LAUNCH_APP = 'LAUNCH_APP';
+export const RETRIEVE_INSTALLED_APPS = 'RETRIEVE_INSTALLED_APPS';
+export const RETRIEVE_INSTALLED_APPS_SUCCESS = 'RETRIEVE_INSTALLED_APPS_SUCCESS';
+export const RETRIEVE_INSTALLED_APPS_ERROR = 'RETRIEVE_INSTALLED_APPS_ERROR';
 
-const initialState = new InitialState();
+function launchAppAction(app) {
+    return {
+        type: LAUNCH_APP,
+        app,
+    };
+}
 
-const reducer = (state = initialState, action) => {
-    switch (action.type) {
-        case NavMenuActions.ITEM_SELECTED:
-            return state.set('selectedItemId', action.id);
-        default:
-            return state;
-    }
-};
+function retrieveInstalledAppsAction() {
+    return {
+        type: RETRIEVE_INSTALLED_APPS,
+    };
+}
 
-export default reducer;
+function retrieveInstalledAppsSuccess(apps) {
+    return {
+        type: RETRIEVE_INSTALLED_APPS_SUCCESS,
+        apps,
+    };
+}
+
+function retrieveInstalledAppsError() {
+    return {
+        type: RETRIEVE_INSTALLED_APPS_ERROR,
+    };
+}
+
+export function retrieveInstalledApps() {
+    return dispatch => {
+        dispatch(retrieveInstalledAppsAction());
+        try {
+            const apps = getInstalledApps();
+            dispatch(retrieveInstalledAppsSuccess(apps));
+        } catch (error) {
+            dispatch(retrieveInstalledAppsError());
+            dispatch(ErrorDialogActions.showDialog(error.message));
+        }
+    };
+}
+
+export function launchApp(app) {
+    return dispatch => {
+        dispatch(launchAppAction(app));
+
+        // The apps in state are Immutable Maps which cannot be sent over IPC.
+        // Converting to plain JS object before sending to the main process.
+        const appObj = app.toJS();
+        ipcRenderer.send('open-app', appObj);
+    };
+}
