+++ lib/windows/launcher/actions/navMenuActions.js
