+++ lib/windows/launcher/components/AppListView.jsx
