+++ lib/windows/launcher/components/AppManagementView.jsx
