+++ lib/windows/launcher/components/MainView.jsx
