+++ lib/windows/launcher/components/SettingsView.jsx
