+++ lib/windows/launcher/containers/AppListViewContainer.js
@@ -49,7 +49,7 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
     return {
         onMount: () => dispatch(AppsActions.retrieveInstalledApps()),
-        onAppSelected: app => dispatch(AppsActions.loadApp(app)),
+        onAppSelected: app => dispatch(AppsActions.launchApp(app)),
     };
 }
