+++ lib/windows/launcher/containers/ErrorDialogContainer.js
