+++ lib/windows/launcher/containers/MainViewContainer.js
