+++ lib/windows/launcher/containers/NavMenuContainer.js
