+++ lib/windows/launcher/containers/RootContainer.js
