+++ lib/windows/launcher/index.html
@@ -7,6 +7,6 @@
 <body>
 <div id='webapp'></div>
 
-<script src="../../../dist/loader-window-bundle.js"></script>
+<script src="../../../dist/launcher-window-bundle.js"></script>
 </body>
 </html>
