+++ lib/windows/launcher/index.js
@@ -34,17 +34,15 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { PropTypes } from 'react';
+import 'babel-polyfill';
 
-const AppLoadButton = ({ onClick }) => (
-    <button title="Load app" className="btn btn-primary core-btn" onClick={onClick}>
-        <span className="icon-play" />
-        <span>Load</span>
-    </button>
-);
+import React from 'react';
+import { render } from 'react-dom';
+import RootContainer from './containers/RootContainer';
+import configureStore from '../../store/configureStore';
+import rootReducer from './reducers';
+import '../../../resources/css/launcher.less';
 
-AppLoadButton.propTypes = {
-    onClick: PropTypes.func.isRequired,
-};
-
-export default AppLoadButton;
+const store = configureStore(rootReducer);
+const rootElement = React.createElement(RootContainer, { store });
+render(rootElement, document.getElementById('webapp'));
