+++ lib/windows/launcher/models/index.js
