+++ lib/windows/launcher/reducers/__tests__/appsReducer-test.js
