+++ lib/windows/launcher/reducers/appsReducer.js
