+++ lib/windows/launcher/reducers/index.js
