+++ lib/windows/launcher/reducers/navMenuReducer.js
@@ -34,67 +34,27 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { mount } from 'enzyme';
-import AppListItem from '../AppListItem';
-import AppLoadButton from '../AppLoadButton';
-import { getImmutableInstalledApp } from '../../models';
+import { Record } from 'immutable';
+import * as NavMenuActions from '../actions/navMenuActions';
 
-const app = getImmutableInstalledApp({
-    name: 'pc-nrfconnect-foobar',
-    version: '1.2.3',
-    path: '/path/to/pc-nrfconnect-foobar',
-    isOfficial: true,
+const InitialState = Record({
+    menuItems: [
+        { id: 1, text: 'Launch app', iconClass: 'icon-ellipsis-vert' },
+        { id: 2, text: 'Add/remove apps', iconClass: 'icon-list-add' },
+        { id: 3, text: 'Settings', iconClass: 'icon-wrench' },
+    ],
+    selectedItemId: 1,
 });
 
-describe('AppListItem', () => {
-    it('should render with required props', () => {
-        expect(renderer.create(
-            <AppListItem
-                app={app}
-                onClick={() => {}}
-            />,
-        )).toMatchSnapshot();
-    });
+const initialState = new InitialState();
 
-    it('should render with displayName instead of name if provided', () => {
-        expect(renderer.create(
-            <AppListItem
-                app={app.set('displayName', 'Foobar Tool')}
-                onClick={() => {}}
-            />,
-        )).toMatchSnapshot();
-    });
+const reducer = (state = initialState, action) => {
+    switch (action.type) {
+        case NavMenuActions.ITEM_SELECTED:
+            return state.set('selectedItemId', action.id);
+        default:
+            return state;
+    }
+};
 
-    it('should render custom icon if provided', () => {
-        expect(renderer.create(
-            <AppListItem
-                app={app.set('iconPath', './path/to/icon.png')}
-                onClick={() => {}}
-            />,
-        )).toMatchSnapshot();
-    });
-
-    it('should render official if isOfficial is true', () => {
-        expect(renderer.create(
-            <AppListItem
-                app={app.set('isOfficial', true)}
-                onClick={() => {}}
-            />,
-        )).toMatchSnapshot();
-    });
-
-    it('should invoke onClick when load button has been clicked', () => {
-        const onClick = jest.fn();
-        const wrapper = mount(
-            <AppListItem
-                app={app}
-                onClick={onClick}
-            />,
-        );
-        wrapper.find(AppLoadButton).first().simulate('click');
-
-        expect(onClick).toHaveBeenCalled();
-    });
-});
+export default reducer;
